+++ lib/shrine/plugins/restore_cached_data.rb
@@ -13,7 +13,7 @@ class Shrine
         def assign_cached(value)
           cached_file = uploaded_file(value) do |cached_file|
             next unless cache.uploaded?(cached_file)
-            real_metadata = cache.extract_metadata(cached_file.to_io, context)
+            real_metadata = cache.extract_metadata(cached_file.tap(&:to_io), context)
             cached_file.metadata.update(real_metadata)
             cached_file.close
           end